- Move all remaining benchmark tests within benchmarks folder
